Fixed DVT compilation error under 64 bit machine

+++ dvt/samples/sm_udp/sm_udp_service.c
@@ -277,7 +277,7 @@ static connector_callback_status_t app_handle_device_request(connector_request_i
         {
             connector_data_service_receive_data_t * const recv_ptr = cb_data;
 
-            APP_DEBUG("Data: \"%.*s\".\n", recv_ptr->bytes_used, (char *)recv_ptr->buffer);
+            APP_DEBUG("Data: \"%.*s\".\n", (int)recv_ptr->bytes_used, (char *)recv_ptr->buffer);
             break;
         }
